feat: Refine mobile filter UI and polish spacing
Refined the filter pills UI based on user feedback to create a cleaner, more polished mobile experience.

**Filter Pills Improvements:**
- Repositioned filters below search controls (not inside search component)
- Filters now span full width and align flush with search input
- Removed visual separator borders (top and bottom lines)
- Tighter, more compact pill styling with improved spacing
- Color-coded pills: blue (title), green (author), purple (subject)
- Polished "Clear All" button with better visual hierarchy
- Removed drag indicator from modal (X button sufficient)

**Layout Changes:**
- Filter pills moved from intelligentSearch.tsx to searchHeader.tsx
- Pills render as sibling to controls row, not nested
- Full-width horizontal scroll without padding constraints
- Border only below search controls, not below filters

**Styling Polish:**
- Reduced pill padding: gap-1.5, px-2.5
- Changed corners from rounded-lg to rounded-md
- Aligned filters to left edge matching search input
- Consistent spacing: py-3 sm:py-4 throughout

**User Feedback Addressed:**
- "Filter UI looks fucking horrible" → Clean, minimal design
- "Still has horizontal line there" → Removed all separator lines
- "Align filters on left with input box" → Flush left alignment

**Files Modified:**
- src/app/library/components/searchHeader.tsx - Filter pills relocated here
- src/app/library/components/intelligentSearch.tsx - Pills removed
- src/components/ui/modal.tsx - Drag indicator removed
