+++ src/app/library/components/intelligentSearch.tsx
@@ -29,16 +29,11 @@ const IntelligentSearch: FC<IntelligentSearchProps> = ({ onClose }) => {
   const dropdownRef = useRef<HTMLDivElement>(null);
   const debounceTimerRef = useRef<NodeJS.Timeout | null>(null);
 
-  // Track active filters from URL params (now supporting comma-separated values)
+  // Track active filters from URL params
   const activeTitle = searchParams.get("title");
   const activeAuthors = searchParams.get("authors");
   const activeSubjects = searchParams.get("subjects");
   const hasActiveFilters = !!(activeTitle || activeAuthors || activeSubjects);
-
-  // Parse comma-separated filter values
-  const activeTitles = activeTitle ? activeTitle.split(",").filter(Boolean) : [];
-  const activeAuthorsList = activeAuthors ? activeAuthors.split(",").filter(Boolean) : [];
-  const activeSubjectsList = activeSubjects ? activeSubjects.split(",").filter(Boolean) : [];
 
   // Fetch suggestions with debouncing
   useEffect(() => {
@@ -321,43 +316,6 @@ const IntelligentSearch: FC<IntelligentSearchProps> = ({ onClose }) => {
     setTimeout(() => setQuery(currentQuery), 0);
   };
 
-  // Handle clear all filters
-  const handleClearAll = () => {
-    router.push("/library");
-    setQuery("");
-    setIsOpen(false);
-    onClose?.();
-  };
-
-  // Handle removing individual filter value
-  const handleRemoveFilter = (
-    filterType: "title" | "authors" | "subjects",
-    value?: string
-  ) => {
-    const params = new URLSearchParams(searchParams);
-    params.delete("page");
-
-    if (value) {
-      // Remove specific value from comma-separated list
-      const currentValue = params.get(filterType);
-      if (currentValue) {
-        const valuesList = currentValue.split(",").filter(Boolean);
-        const updatedList = valuesList.filter((v) => v !== value);
-
-        if (updatedList.length > 0) {
-          params.set(filterType, updatedList.join(","));
-        } else {
-          params.delete(filterType);
-        }
-      }
-    } else {
-      // Remove entire filter type
-      params.delete(filterType);
-    }
-
-    router.push(`/library/?${params.toString()}`);
-  };
-
   // Render suggestion sections
   const renderSection = (
     title: string,
@@ -397,74 +355,7 @@ const IntelligentSearch: FC<IntelligentSearchProps> = ({ onClose }) => {
   const hasResults = totalItems > 0;
 
   return (
-    <div className="w-full space-y-3">
-      {/* Active Filter Pills - Now Above Search Input */}
-      {hasActiveFilters && (
-        <div className="bg-zinc-900/30 border border-zinc-800/50 rounded-lg p-3">
-          <div className="flex items-center justify-between gap-3 mb-2">
-            <span className="text-xs font-semibold text-zinc-400 uppercase tracking-wider">
-              Active Filters
-            </span>
-            <button
-              onClick={handleClearAll}
-              className="px-2.5 py-1 text-xs font-medium text-zinc-400 hover:text-zinc-200 bg-zinc-900/50 hover:bg-zinc-800/50 border border-zinc-800 hover:border-zinc-700 rounded transition-all duration-150"
-            >
-              Clear All
-            </button>
-          </div>
-          <div className="flex flex-nowrap md:flex-wrap gap-2 overflow-x-auto pb-2 -mx-3 px-3 md:mx-0 md:px-0 hide-scrollbar">
-            {activeTitles.map((title) => (
-              <div
-                key={`title-${title}`}
-                className="inline-flex items-center gap-2 px-3 py-1.5 bg-zinc-700/50 border border-zinc-600/50 rounded-lg text-xs flex-shrink-0"
-              >
-                <span className="font-semibold text-zinc-300">Title:</span>
-                <span className="text-zinc-100 truncate max-w-[100px] sm:max-w-[200px]">{title}</span>
-                <button
-                  onClick={() => handleRemoveFilter("title", title)}
-                  className="ml-1 p-1.5 sm:p-1 min-w-[28px] min-h-[28px] text-zinc-400 hover:text-zinc-100 transition-colors duration-150 rounded hover:bg-zinc-600/30"
-                  aria-label={`Remove title filter: ${title}`}
-                >
-                  <X className="w-4 h-4 sm:w-3.5 sm:h-3.5" />
-                </button>
-              </div>
-            ))}
-            {activeAuthorsList.map((author) => (
-              <div
-                key={`author-${author}`}
-                className="inline-flex items-center gap-2 px-3 py-1.5 bg-blue-900/50 border border-blue-800/50 rounded-lg text-xs flex-shrink-0"
-              >
-                <span className="font-semibold text-blue-200">Author:</span>
-                <span className="text-blue-100 truncate max-w-[100px] sm:max-w-[200px]">{author}</span>
-                <button
-                  onClick={() => handleRemoveFilter("authors", author)}
-                  className="ml-1 p-1.5 sm:p-1 min-w-[28px] min-h-[28px] text-blue-300 hover:text-blue-100 transition-colors duration-150 rounded hover:bg-blue-800/30"
-                  aria-label={`Remove author filter: ${author}`}
-                >
-                  <X className="w-4 h-4 sm:w-3.5 sm:h-3.5" />
-                </button>
-              </div>
-            ))}
-            {activeSubjectsList.map((subject) => (
-              <div
-                key={`subject-${subject}`}
-                className="inline-flex items-center gap-2 px-3 py-1.5 bg-purple-900/50 border border-purple-800/50 rounded-lg text-xs flex-shrink-0"
-              >
-                <span className="font-semibold text-purple-200">Subject:</span>
-                <span className="text-purple-100 truncate max-w-[100px] sm:max-w-[200px]">{subject}</span>
-                <button
-                  onClick={() => handleRemoveFilter("subjects", subject)}
-                  className="ml-1 p-1.5 sm:p-1 min-w-[28px] min-h-[28px] text-purple-300 hover:text-purple-100 transition-colors duration-150 rounded hover:bg-purple-800/30"
-                  aria-label={`Remove subject filter: ${subject}`}
-                >
-                  <X className="w-4 h-4 sm:w-3.5 sm:h-3.5" />
-                </button>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
+    <div className="w-full">
       <div className="relative w-full">
         {/* Search Input */}
         <div className="relative">
